fixing the part display in the context menu; fixing the file name cases on Windows;

+++ branches/1.2/olex/main_gui.cpp
@@ -550,7 +550,7 @@ void TMainForm::ObjectUnderMouse(AGDrawObject *G)  {
     else {
       pmAtom->Check(ID_AtomPartChangeLast, true);
       pmAtom->SetLabel(ID_AtomPartChangeLast, (olxstr("Custom: ") <<
-        XA->CAtom().GetPart()).u_str());
+        (int)XA->CAtom().GetPart()).u_str());
     }
     // Uiso
     if (XA->CAtom().GetEllipsoid() == NULL) {
